+++ sdk/servicenetworking/azure-resourcemanager-servicenetworking/src/main/java/com/azure/resourcemanager/servicenetworking/models/ActionType.java
@@ -8,14 +8,18 @@ import com.azure.core.util.ExpandableStringEnum;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import java.util.Collection;
 
-/** Enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs. */
+/**
+ * Enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs.
+ */
 public final class ActionType extends ExpandableStringEnum<ActionType> {
-    /** Static value Internal for ActionType. */
+    /**
+     * Static value Internal for ActionType.
+     */
     public static final ActionType INTERNAL = fromString("Internal");
 
     /**
      * Creates a new instance of ActionType value.
-     *
+     * 
      * @deprecated Use the {@link #fromString(String)} factory method.
      */
     @Deprecated
@@ -24,7 +28,7 @@ public final class ActionType extends ExpandableStringEnum<ActionType> {
 
     /**
      * Creates or finds a ActionType from its string representation.
-     *
+     * 
      * @param name a name to look for.
      * @return the corresponding ActionType.
      */
@@ -35,7 +39,7 @@ public final class ActionType extends ExpandableStringEnum<ActionType> {
 
     /**
      * Gets known ActionType values.
-     *
+     * 
      * @return known ActionType values.
      */
     public static Collection<ActionType> values() {