+++ sdk/communication/azure-communication-email/CHANGELOG.md
@@ -11,6 +11,16 @@
 - An `EmailMessage` with null recipient addresses can no longer be sent.
 ### Other Changes
 
+## 1.0.9 (2023-12-04)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core-http-netty` from `1.13.10` to version `1.13.11`.
+- Upgraded `azure-communication-common` from `1.2.14` to version `1.2.15`.
+- Upgraded `azure-core` from `1.45.0` to version `1.45.1`.
+
 ## 1.0.8 (2023-11-20)
 
 ### Other Changes