+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### Other Changes
 
+## 12.24.1 (2023-12-04)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-storage-blob` from `12.25.0` to version `12.25.1`.
+- Upgraded `azure-core-http-netty` from `1.13.10` to version `1.13.11`.
+- Upgraded `azure-core` from `1.45.0` to version `1.45.1`.
+
 ## 12.24.0 (2023-11-08)
 
 ### Features Added